+++ src/i18n/i18n.ts
@@ -5,6 +5,7 @@ import { getLanguage } from 'obsidian';
 import be from './locales/be.json';
 import de from './locales/de.json';
 import en from './locales/en.json';
+import es from './locales/es.json';
 import ko from './locales/ko.json';
 import pt_br from './locales/pt_br.json';
 import ru from './locales/ru.json';
@@ -36,6 +37,7 @@ export const initializeI18n = async () => {
                 be: { translation: be }, // Belarusian
                 de: { translation: de }, // German
                 en: { translation: en }, // English
+                es: { translation: es }, // Spanish
                 ko: { translation: ko }, // Korean
                 'pt-BR': { translation: pt_br }, // Portuguese (Brazil)
                 ru: { translation: ru }, // Russian